limiting encrypted image pull to linux for now
